@@ -4,15 +4,18 @@
   "private": true,
   "proxy": "http://localhost:4000",
   "dependencies": {
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/user-event": "^13.5.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.34",
     "@fortawesome/free-solid-svg-icons": "^5.15.2",
     "@fortawesome/react-fontawesome": "^0.1.14",
     "axios": "^0.21.1",
     "concurrently": "^6.0.0",
-    "react": "16.4.1",
-    "react-dom": "16.4.1",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "1.1.4",
+    "react-scripts": "5.0.1",
     "react-toastify": "4.5.2"
   },
   "scripts": {
